@@ -3,19 +3,22 @@ package toyhose
 import (
 	"context"
 	"fmt"
-	"io/ioutil"
+	"io"
+	"os"
 	"path/filepath"
 	"regexp"
 	"testing"
 	"time"
 
-	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/service/firehose"
-	"github.com/aws/aws-sdk-go/service/s3"
+	"github.com/aws/aws-sdk-go-v2/aws"
+	fhtypes "github.com/aws/aws-sdk-go-v2/service/firehose/types"
+	"github.com/aws/aws-sdk-go-v2/service/s3"
+	s3types "github.com/aws/aws-sdk-go-v2/service/s3/types"
 	"github.com/google/uuid"
 )
 
 func TestS3DestinationForExceededDataSize(t *testing.T) {
+	awsConf := awsConfig(t)
 	s3cli := s3Client(awsConf, s3EndpointURL)
 	bucketName := "store-s3-test-" + uuid.New().String()
 	if err := setupS3(t, s3cli, bucketName); err != nil {
@@ -31,9 +34,9 @@ func TestS3DestinationForExceededDataSize(t *testing.T) {
 		deliveryName: "foobar",
 		prefix:       aws.String("aaa"),
 		bucketARN:    "arn:aws:s3:::" + bucketName,
-		bufferingHints: &firehose.BufferingHints{
-			SizeInMBs:         aws.Int64(1),
-			IntervalInSeconds: aws.Int64(60),
+		bufferingHints: &fhtypes.BufferingHints{
+			SizeInMBs:         aws.Int32(1),
+			IntervalInSeconds: aws.Int32(60),
 		},
 	}
 	ctx, cancel := context.WithCancel(context.Background())
@@ -44,7 +47,7 @@ func TestS3DestinationForExceededDataSize(t *testing.T) {
 	}
 	go dst.Run(ctx, conf, ch)
 
-	b, err := ioutil.ReadFile(filepath.Join("testdata", "dummy.json"))
+	b, err := os.ReadFile(filepath.Join("testdata", "dummy.json"))
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -55,9 +58,9 @@ func TestS3DestinationForExceededDataSize(t *testing.T) {
 		ch <- newDeliveryRecord(b)
 		sent += len(b)
 	}
-	var obj *s3.Object
+	var obj s3types.Object
 	for i := 0; i < 20; i++ {
-		out, err := s3cli.ListObjects(&s3.ListObjectsInput{
+		out, err := s3cli.ListObjects(ctx, &s3.ListObjectsInput{
 			Bucket: &bucketName,
 			Prefix: dst.prefix,
 		})
@@ -70,17 +73,17 @@ func TestS3DestinationForExceededDataSize(t *testing.T) {
 		}
 		time.Sleep(100 * time.Millisecond)
 	}
-	if obj == nil {
+	if obj.Key == nil {
 		t.Fatal("s3 object not found")
 	}
-	out, err := s3cli.GetObject(&s3.GetObjectInput{
+	out, err := s3cli.GetObject(ctx, &s3.GetObjectInput{
 		Bucket: &bucketName,
 		Key:    obj.Key,
 	})
 	if err != nil {
 		t.Fatal(err)
 	}
-	content, err := ioutil.ReadAll(out.Body)
+	content, err := io.ReadAll(out.Body)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -90,6 +93,7 @@ func TestS3DestinationForExceededDataSize(t *testing.T) {
 }
 
 func TestS3DestinationForExceededInterval(t *testing.T) {
+	awsConf := awsConfig(t)
 	s3cli := s3Client(awsConf, s3EndpointURL)
 	bucketName := "store-s3-test-" + uuid.New().String()
 	if err := setupS3(t, s3cli, bucketName); err != nil {
@@ -104,9 +108,9 @@ func TestS3DestinationForExceededInterval(t *testing.T) {
 		},
 		deliveryName: "foobar",
 		bucketARN:    "arn:aws:s3:::" + bucketName,
-		bufferingHints: &firehose.BufferingHints{
-			SizeInMBs:         aws.Int64(50),
-			IntervalInSeconds: aws.Int64(1), // for test use only
+		bufferingHints: &fhtypes.BufferingHints{
+			SizeInMBs:         aws.Int32(50),
+			IntervalInSeconds: aws.Int32(1), // for test use only
 		},
 		prefix: aws.String("bbb"),
 	}
@@ -118,7 +122,7 @@ func TestS3DestinationForExceededInterval(t *testing.T) {
 	}
 	go dst.Run(ctx, conf, ch)
 
-	b, err := ioutil.ReadFile(filepath.Join("testdata", "dummy.json"))
+	b, err := os.ReadFile(filepath.Join("testdata", "dummy.json"))
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -127,9 +131,9 @@ func TestS3DestinationForExceededInterval(t *testing.T) {
 		ch <- newDeliveryRecord(b)
 		time.Sleep(800 * time.Millisecond)
 	}
-	var objects []*s3.Object
+	var objects []s3types.Object
 	for i := 0; i < 50; i++ {
-		out, err := s3cli.ListObjects(&s3.ListObjectsInput{
+		out, err := s3cli.ListObjects(ctx, &s3.ListObjectsInput{
 			Bucket: &bucketName,
 			Prefix: dst.prefix,
 		})
@@ -146,14 +150,14 @@ func TestS3DestinationForExceededInterval(t *testing.T) {
 		t.Fatal("s3 objects not found")
 	}
 	for _, c := range objects {
-		out, err := s3cli.GetObject(&s3.GetObjectInput{
+		out, err := s3cli.GetObject(ctx, &s3.GetObjectInput{
 			Bucket: &bucketName,
 			Key:    c.Key,
 		})
 		if err != nil {
 			t.Fatal(err)
 		}
-		content, err := ioutil.ReadAll(out.Body)
+		content, err := io.ReadAll(out.Body)
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -168,6 +172,7 @@ func TestS3DestinationForExceededInterval(t *testing.T) {
 var re = regexp.MustCompile("\\d$")
 
 func TestS3DestinationWithDisableBuffering(t *testing.T) {
+	awsConf := awsConfig(t)
 	s3cli := s3Client(awsConf, s3EndpointURL)
 	bucketName := "store-s3-test-" + uuid.New().String()
 	if err := setupS3(t, s3cli, bucketName); err != nil {
@@ -193,7 +198,7 @@ func TestS3DestinationWithDisableBuffering(t *testing.T) {
 	}
 	go dst.Run(ctx, conf, ch)
 
-	b, err := ioutil.ReadFile(filepath.Join("testdata", "dummy.json"))
+	b, err := os.ReadFile(filepath.Join("testdata", "dummy.json"))
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -201,9 +206,9 @@ func TestS3DestinationWithDisableBuffering(t *testing.T) {
 	for i := 0; i < 5; i++ {
 		ch <- newDeliveryRecord(append(b, []byte(fmt.Sprint(i))...))
 	}
-	var objects []*s3.Object
+	var objects []s3types.Object
 	for i := 0; i < 20; i++ {
-		out, err := s3cli.ListObjects(&s3.ListObjectsInput{
+		out, err := s3cli.ListObjects(ctx, &s3.ListObjectsInput{
 			Bucket: &bucketName,
 			Prefix: dst.prefix,
 		})
@@ -220,14 +225,14 @@ func TestS3DestinationWithDisableBuffering(t *testing.T) {
 		t.Fatal("s3 objects not found")
 	}
 	for _, obj := range objects {
-		out, err := s3cli.GetObject(&s3.GetObjectInput{
+		out, err := s3cli.GetObject(ctx, &s3.GetObjectInput{
 			Bucket: &bucketName,
 			Key:    obj.Key,
 		})
 		if err != nil {
 			t.Fatal(err)
 		}
-		content, err := ioutil.ReadAll(out.Body)
+		content, err := io.ReadAll(out.Body)
 		if err != nil {
 			t.Fatal(err)
 		}
