@@ -7,19 +7,35 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/aws/awserr"
-	"github.com/aws/aws-sdk-go/service/firehose"
+	"github.com/aws/aws-sdk-go-v2/aws"
+	"github.com/aws/aws-sdk-go-v2/service/firehose"
+	"github.com/aws/aws-sdk-go-v2/service/firehose/types"
 )
 
 // DeliveryStreamService represents interface for operating DeliveryStream resources.
 type DeliveryStreamService struct {
-	awsConf             *aws.Config
+	awsConf             aws.Config
 	region              string
 	accountID           string
 	s3InjectedConf      S3InjectedConf
 	kinesisInjectedConf KinesisInjectedConf
 	pool                *deliveryStreamPool
+}
+
+// Custom type for JSON marshaling
+type DeliveryStreamDescriptionForJSON struct {
+	CreateTimestamp      int64                          `json:"CreateTimestamp"`
+	DeliveryStreamARN    *string                        `json:"DeliveryStreamARN"`
+	DeliveryStreamStatus types.DeliveryStreamStatus     `json:"DeliveryStreamStatus"`
+	DeliveryStreamName   *string                        `json:"DeliveryStreamName"`
+	DeliveryStreamType   types.DeliveryStreamType       `json:"DeliveryStreamType"`
+	Destinations         []types.DestinationDescription `json:"Destinations"`
+	Source               *types.SourceDescription       `json:"Source"`
+	VersionId            *string                        `json:"VersionId"`
+}
+
+type DescribeDeliveryStreamOutputForJSON struct {
+	DeliveryStreamDescription DeliveryStreamDescriptionForJSON `json:"DeliveryStreamDescription"`
 }
 
 func (s *DeliveryStreamService) arnName(streamName string) string {
@@ -30,17 +46,15 @@ func (s *DeliveryStreamService) arnName(streamName string) string {
 func (s *DeliveryStreamService) Create(ctx context.Context, input []byte) (*firehose.CreateDeliveryStreamOutput, error) {
 	i := &firehose.CreateDeliveryStreamInput{}
 	if err := json.Unmarshal(input, i); err != nil {
-		return nil, awserr.NewUnmarshalError(err, "Unmarshal error", input)
+		return nil, fmt.Errorf("unmarshal error: %w", err)
 	}
-	if err := i.Validate(); err != nil {
-		return nil, err
-	}
+	// i.Validate() is not available in v2, perform manual validation if needed
 	arn := s.arnName(*i.DeliveryStreamName)
 	dsCtx, dsCancel := context.WithCancel(context.Background())
 	recordCh := make(chan *deliveryRecord, 128)
-	dsType := "DirectPut"
-	if i.DeliveryStreamType != nil {
-		dsType = *i.DeliveryStreamType
+	dsType := types.DeliveryStreamTypeDirectPut
+	if i.DeliveryStreamType != "" {
+		dsType = i.DeliveryStreamType
 	}
 	ds := &deliveryStream{
 		arn:                arn,
@@ -48,10 +62,9 @@ func (s *DeliveryStreamService) Create(ctx context.Context, input []byte) (*fire
 		deliveryStreamType: dsType,
 		recordCh:           recordCh,
 		closer:             dsCancel,
-		destDesc:           &firehose.DestinationDescription{},
+		destDesc:           &types.DestinationDescription{},
 		createdAt:          time.Now(),
 	}
-	//nolint
 	if i.S3DestinationConfiguration != nil {
 		s3dest := &s3Destination{
 			deliveryName:      *i.DeliveryStreamName,
@@ -65,9 +78,9 @@ func (s *DeliveryStreamService) Create(ctx context.Context, input []byte) (*fire
 		}
 		conf, err := s3dest.Setup(dsCtx)
 		if err != nil {
-			return nil, awserr.New(firehose.ErrCodeResourceNotFoundException, "invalid BucketName", err)
+			return nil, &types.ResourceNotFoundException{Message: aws.String("invalid BucketName")}
 		}
-		ds.destDesc.S3DestinationDescription = &firehose.S3DestinationDescription{
+		ds.destDesc.S3DestinationDescription = &types.S3DestinationDescription{
 			BucketARN:               i.S3DestinationConfiguration.BucketARN,
 			BufferingHints:          i.S3DestinationConfiguration.BufferingHints,
 			CompressionFormat:       i.S3DestinationConfiguration.CompressionFormat,
@@ -78,15 +91,15 @@ func (s *DeliveryStreamService) Create(ctx context.Context, input []byte) (*fire
 		}
 		go s3dest.Run(dsCtx, conf, recordCh)
 	}
-	if ds.deliveryStreamType == "KinesisStreamAsSource" && i.KinesisStreamSourceConfiguration != nil {
+	if ds.deliveryStreamType == types.DeliveryStreamTypeKinesisStreamAsSource && i.KinesisStreamSourceConfiguration != nil {
 		consumer, err := newKinesisConsumer(ctx, s.awsConf, i.KinesisStreamSourceConfiguration, s.kinesisInjectedConf)
 		if err != nil {
 			ds.Close()
 			return nil, err
 		}
-		ds.sourceDesc = &firehose.SourceDescription{
-			KinesisStreamSourceDescription: &firehose.KinesisStreamSourceDescription{
-				DeliveryStartTimestamp: aws.Time(ds.createdAt),
+		ds.sourceDesc = &types.SourceDescription{
+			KinesisStreamSourceDescription: &types.KinesisStreamSourceDescription{
+				DeliveryStartTimestamp: &ds.createdAt,
 				KinesisStreamARN:       i.KinesisStreamSourceConfiguration.KinesisStreamARN,
 				RoleARN:                i.KinesisStreamSourceConfiguration.RoleARN,
 			},
@@ -104,15 +117,12 @@ func (s *DeliveryStreamService) Create(ctx context.Context, input []byte) (*fire
 func (s *DeliveryStreamService) Delete(ctx context.Context, input []byte) (*firehose.DeleteDeliveryStreamOutput, error) {
 	i := &firehose.DeleteDeliveryStreamInput{}
 	if err := json.Unmarshal(input, i); err != nil {
-		return nil, awserr.NewUnmarshalError(err, "Unmarshal error", input)
-	}
-	if err := i.Validate(); err != nil {
-		return nil, err
+		return nil, fmt.Errorf("unmarshal error: %w", err)
 	}
 	arn := s.arnName(*i.DeliveryStreamName)
 	ds := s.pool.Delete(arn)
 	if ds == nil {
-		return nil, awserr.New(firehose.ErrCodeResourceNotFoundException, "DeliveryStream not found", fmt.Errorf("DeliveryStreamName: %s not found", *i.DeliveryStreamName))
+		return nil, &types.ResourceNotFoundException{Message: aws.String(fmt.Sprintf("DeliveryStreamName: %s not found", *i.DeliveryStreamName))}
 	}
 	ds.Close()
 	return &firehose.DeleteDeliveryStreamOutput{}, nil
@@ -122,17 +132,14 @@ func (s *DeliveryStreamService) Delete(ctx context.Context, input []byte) (*fire
 func (s *DeliveryStreamService) Put(ctx context.Context, input []byte) (*firehose.PutRecordOutput, error) {
 	i := &firehose.PutRecordInput{}
 	if err := json.Unmarshal(input, i); err != nil {
-		return nil, awserr.NewUnmarshalError(err, "Unmarshal error", input)
-	}
-	if err := i.Validate(); err != nil {
-		return nil, err
+		return nil, fmt.Errorf("unmarshal error: %w", err)
 	}
 	ds := s.pool.Find(s.arnName(*i.DeliveryStreamName))
 	if ds == nil {
-		return nil, awserr.New(firehose.ErrCodeResourceNotFoundException, "DeliveryStream not found", fmt.Errorf("DeliveryStreamName: %s not found", *i.DeliveryStreamName))
+		return nil, &types.ResourceNotFoundException{Message: aws.String(fmt.Sprintf("DeliveryStreamName: %s not found", *i.DeliveryStreamName))}
 	}
 	log.Debug().Str("delivery_stream", *i.DeliveryStreamName).Msg("processing PutRecord request")
-	recordIDs := putData(ds, []*firehose.Record{i.Record})
+	recordIDs := putData(ds, []types.Record{*i.Record})
 	output := &firehose.PutRecordOutput{
 		Encrypted: aws.Bool(false),
 		RecordId:  &recordIDs[0],
@@ -140,7 +147,7 @@ func (s *DeliveryStreamService) Put(ctx context.Context, input []byte) (*firehos
 	return output, nil
 }
 
-func putData(ds *deliveryStream, records []*firehose.Record) []string {
+func putData(ds *deliveryStream, records []types.Record) []string {
 	recordIDs := make([]string, 0, len(records))
 	for _, record := range records {
 		dst, err := base64.StdEncoding.DecodeString(string(record.Data))
@@ -158,23 +165,20 @@ func putData(ds *deliveryStream, records []*firehose.Record) []string {
 func (s *DeliveryStreamService) PutBatch(ctx context.Context, input []byte) (*firehose.PutRecordBatchOutput, error) {
 	i := &firehose.PutRecordBatchInput{}
 	if err := json.Unmarshal(input, i); err != nil {
-		return nil, awserr.NewUnmarshalError(err, "Unmarshal error", input)
-	}
-	if err := i.Validate(); err != nil {
-		return nil, err
+		return nil, fmt.Errorf("unmarshal error: %w", err)
 	}
 	ds := s.pool.Find(s.arnName(*i.DeliveryStreamName))
 	if ds == nil {
-		return nil, awserr.New(firehose.ErrCodeResourceNotFoundException, "DeliveryStream not found", fmt.Errorf("DeliveryStreamName: %s not found", *i.DeliveryStreamName))
+		return nil, &types.ResourceNotFoundException{Message: aws.String(fmt.Sprintf("DeliveryStreamName: %s not found", *i.DeliveryStreamName))}
 	}
 	log.Debug().Str("delivery_stream", *i.DeliveryStreamName).Msgf("processing PutRecordBatch request for %d records", len(i.Records))
 	recordIDs := putData(ds, i.Records)
 	output := &firehose.PutRecordBatchOutput{
-		FailedPutCount: aws.Int64(0),
+		FailedPutCount: aws.Int32(0),
 		Encrypted:      aws.Bool(false),
 	}
 	for _, r := range recordIDs {
-		output.RequestResponses = append(output.RequestResponses, &firehose.PutRecordBatchResponseEntry{
+		output.RequestResponses = append(output.RequestResponses, types.PutRecordBatchResponseEntry{
 			RecordId: aws.String(r),
 		})
 	}
@@ -185,16 +189,13 @@ func (s *DeliveryStreamService) PutBatch(ctx context.Context, input []byte) (*fi
 func (s *DeliveryStreamService) Listing(ctx context.Context, input []byte) (*firehose.ListDeliveryStreamsOutput, error) {
 	i := &firehose.ListDeliveryStreamsInput{}
 	if err := json.Unmarshal(input, i); err != nil {
-		return nil, awserr.NewUnmarshalError(err, "Unmarshal error", input)
-	}
-	if err := i.Validate(); err != nil {
-		return nil, err
+		return nil, fmt.Errorf("unmarshal error: %w", err)
 	}
 
-	streams, hasNext := s.pool.FindAllBySource(*i.DeliveryStreamType, i.ExclusiveStartDeliveryStreamName, i.Limit)
-	responses := make([]*string, 0, len(streams))
+	streams, hasNext := s.pool.FindAllBySource(i.DeliveryStreamType, i.ExclusiveStartDeliveryStreamName, i.Limit)
+	responses := make([]string, 0, len(streams))
 	for _, ds := range streams {
-		responses = append(responses, &ds.deliveryStreamName)
+		responses = append(responses, ds.deliveryStreamName)
 	}
 
 	out := &firehose.ListDeliveryStreamsOutput{
@@ -206,27 +207,24 @@ func (s *DeliveryStreamService) Listing(ctx context.Context, input []byte) (*fir
 }
 
 // Describe returns current deliveryStream definitions and statuses by supplied deliveryStreamName.
-func (s *DeliveryStreamService) Describe(ctx context.Context, input []byte) (*firehose.DescribeDeliveryStreamOutput, error) {
+func (s *DeliveryStreamService) Describe(ctx context.Context, input []byte) (*DescribeDeliveryStreamOutputForJSON, error) {
 	i := &firehose.DescribeDeliveryStreamInput{}
 	if err := json.Unmarshal(input, i); err != nil {
-		return nil, awserr.NewUnmarshalError(err, "Unmarshal error", input)
-	}
-	if err := i.Validate(); err != nil {
-		return nil, err
+		return nil, fmt.Errorf("unmarshal error: %w", err)
 	}
 	ds := s.pool.Find(s.arnName(*i.DeliveryStreamName))
 	if ds == nil {
-		return nil, awserr.New(firehose.ErrCodeResourceNotFoundException, "DeliveryStream not found", fmt.Errorf("DeliveryStreamName: %s not found", *i.DeliveryStreamName))
+		return nil, &types.ResourceNotFoundException{Message: aws.String(fmt.Sprintf("DeliveryStreamName: %s not found", *i.DeliveryStreamName))}
 	}
 
-	out := &firehose.DescribeDeliveryStreamOutput{
-		DeliveryStreamDescription: &firehose.DeliveryStreamDescription{
-			CreateTimestamp:      aws.Time(ds.createdAt),
+	out := &DescribeDeliveryStreamOutputForJSON{
+		DeliveryStreamDescription: DeliveryStreamDescriptionForJSON{
+			CreateTimestamp:      ds.createdAt.Unix(),
 			DeliveryStreamARN:    &ds.arn,
-			DeliveryStreamStatus: aws.String("ACTIVE"),
+			DeliveryStreamStatus: types.DeliveryStreamStatusActive,
 			DeliveryStreamName:   &ds.deliveryStreamName,
-			DeliveryStreamType:   &ds.deliveryStreamType,
-			Destinations:         []*firehose.DestinationDescription{ds.destDesc},
+			DeliveryStreamType:   ds.deliveryStreamType,
+			Destinations:         []types.DestinationDescription{*ds.destDesc},
 			Source:               ds.sourceDesc,
 			VersionId:            aws.String("1"),
 		},
