@@ -1,16 +1,14 @@
 package toyhose
 
 import (
-	"bytes"
 	"encoding/json"
 	"errors"
 	"io"
 	"net/http"
 	"strings"
 
-	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/aws/awserr"
-	"github.com/aws/aws-sdk-go/service/firehose"
+	"github.com/aws/aws-sdk-go-v2/aws"
+	"github.com/aws/aws-sdk-go-v2/service/firehose/types"
 	"github.com/google/uuid"
 )
 
@@ -18,7 +16,7 @@ import (
 type DispatcherConfig struct {
 	S3InjectedConf      S3InjectedConf
 	KinesisInjectedConf KinesisInjectedConf
-	AWSConf             *aws.Config
+	AWSConf             aws.Config
 }
 
 // S3InjectedConf represents injection to S3 destination BufferingHints forcely.
@@ -36,11 +34,10 @@ type KinesisInjectedConf struct {
 
 // NewDispatcher returns Dispatcher object.
 func NewDispatcher(conf *DispatcherConfig) *Dispatcher {
-	cred, _ := conf.AWSConf.Credentials.Get()
 	return &Dispatcher{
 		conf:                conf.AWSConf,
-		accountID:           cred.AccessKeyID,
-		region:              *conf.AWSConf.Region,
+		accountID:           "", // FIXME: Get AccountID from STS or other means if needed.
+		region:              conf.AWSConf.Region,
 		s3InjectedConf:      conf.S3InjectedConf,
 		kinesisInjectedConf: conf.KinesisInjectedConf,
 		pool: &deliveryStreamPool{
@@ -51,7 +48,7 @@ func NewDispatcher(conf *DispatcherConfig) *Dispatcher {
 
 // Dispatcher represents firehose API handler.
 type Dispatcher struct {
-	conf                *aws.Config
+	conf                aws.Config
 	accountID           string
 	region              string
 	s3InjectedConf      S3InjectedConf
@@ -61,6 +58,7 @@ type Dispatcher struct {
 
 // Dispatch handlers HTTP request as http.HandlerFunc interface.
 func (d *Dispatcher) Dispatch(w http.ResponseWriter, r *http.Request) {
+	ctx := r.Context()
 	reqID := uuid.New().String()
 	w.Header().Add("x-amzn-RequestId", reqID)
 	w.Header().Add("Content-Type", r.Header.Get("Content-Type"))
@@ -69,12 +67,12 @@ func (d *Dispatcher) Dispatch(w http.ResponseWriter, r *http.Request) {
 		outputForJSON(w, nil, err)
 		return
 	}
-	b := bytes.NewBuffer([]byte{})
-	if _, err := io.Copy(b, r.Body); err != nil {
+	bodyBytes, err := io.ReadAll(r.Body)
+	if err != nil {
 		outputForJSON(w, nil, err)
 		return
 	}
-	if err := verifyV4(d.conf, r, bytes.NewReader(b.Bytes())); err != nil {
+	if err := verifyV4(ctx, d.conf, r, bodyBytes); err != nil {
 		outputForJSON(w, nil, err)
 		return
 	}
@@ -88,26 +86,27 @@ func (d *Dispatcher) Dispatch(w http.ResponseWriter, r *http.Request) {
 	}
 	switch op {
 	case "CreateDeliveryStream":
-		out, err := svc.Create(r.Context(), b.Bytes())
+		out, err := svc.Create(ctx, bodyBytes)
 		outputForJSON(w, out, err)
 	case "DeleteDeliveryStream":
-		out, err := svc.Delete(r.Context(), b.Bytes())
+		out, err := svc.Delete(ctx, bodyBytes)
 		outputForJSON(w, out, err)
 	case "PutRecord":
-		out, err := svc.Put(r.Context(), b.Bytes())
+		out, err := svc.Put(ctx, bodyBytes)
 		outputForJSON(w, out, err)
 	case "PutRecordBatch":
-		out, err := svc.PutBatch(r.Context(), b.Bytes())
+		out, err := svc.PutBatch(ctx, bodyBytes)
 		outputForJSON(w, out, err)
 	case "ListDeliveryStreams":
-		out, err := svc.Listing(r.Context(), b.Bytes())
+		out, err := svc.Listing(ctx, bodyBytes)
 		outputForJSON(w, out, err)
 	case "DescribeDeliveryStream":
-		out, err := svc.Describe(r.Context(), b.Bytes())
+		// Use custom struct for JSON marshaling
+		out, err := svc.Describe(ctx, bodyBytes)
 		outputForJSON(w, out, err)
 	default:
-		err := awserr.New("InvalidAction", "invalid action received", errInvalidTargetHeader)
-		outputForJSON(w, nil, err)
+		// Consider creating a new error type for v2 or using a generic one
+		outputForJSON(w, nil, errors.New("InvalidAction: invalid action received"))
 	}
 }
 
@@ -116,29 +115,30 @@ type outputSerializable interface {
 	GoString() string
 }
 
-func outputForJSON(w http.ResponseWriter, out outputSerializable, err error) {
+func outputForJSON(w http.ResponseWriter, out interface{}, err error) {
 	if err == nil {
 		if err := json.NewEncoder(w).Encode(out); err != nil {
 			log.Error().Err(err).Msg("failed to decode output json")
 		}
 		return
 	}
-	switch e := err.(type) {
-	case awserr.Error:
-		switch e.Code() {
-		case "InternalFailure":
-			w.WriteHeader(http.StatusInternalServerError)
-		case firehose.ErrCodeServiceUnavailableException:
-			w.WriteHeader(http.StatusServiceUnavailable)
-		case firehose.ErrCodeResourceNotFoundException:
-			w.WriteHeader(http.StatusNotFound)
-		default:
-			w.WriteHeader(http.StatusBadRequest)
-		}
-	default:
+
+	// Simplified error handling for now.
+	// TODO: Implement proper error type checking for v2 SDK errors.
+	var resourceNotFoundException *types.ResourceNotFoundException
+	if errors.As(err, &resourceNotFoundException) {
+		w.WriteHeader(http.StatusNotFound)
+	} else {
 		w.WriteHeader(http.StatusBadRequest)
 	}
-	if err2 := json.NewEncoder(w).Encode(err); err2 != nil {
+
+	// Create a serializable error message
+	errMsg := struct {
+		Message string `json:"message"`
+	}{
+		Message: err.Error(),
+	}
+	if err2 := json.NewEncoder(w).Encode(errMsg); err2 != nil {
 		log.Error().Err(err2).Msg("failed to decode output error json")
 	}
 }
@@ -149,7 +149,7 @@ func parseTarget(tgt string) (string, error) {
 	// ex) Firehose_20150804.CreateDeliveryStream
 	parts := strings.Split(tgt, ".")
 	if len(parts) != 2 || !strings.HasPrefix(parts[0], "Firehose_") {
-		return "", awserr.New("MissingAction", "no action received", errInvalidTargetHeader)
+		return "", errors.New("MissingAction: no action received")
 	}
 	return parts[1], nil
 }
