@@ -8,11 +8,10 @@ import (
 	"sync"
 	"time"
 
-	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/aws/awserr"
-	"github.com/aws/aws-sdk-go/aws/session"
-	"github.com/aws/aws-sdk-go/service/firehose"
-	"github.com/aws/aws-sdk-go/service/kinesis"
+	"github.com/aws/aws-sdk-go-v2/aws"
+	fhtypes "github.com/aws/aws-sdk-go-v2/service/firehose/types"
+	"github.com/aws/aws-sdk-go-v2/service/kinesis"
+	"github.com/aws/aws-sdk-go-v2/service/kinesis/types"
 )
 
 var (
@@ -20,52 +19,55 @@ var (
 	errInvalidStreamARN = errors.New("invalid arn")
 )
 
-func getStreamNameFromARN(conf *aws.Config, arn string) (string, error) {
+func getStreamNameFromARN(conf aws.Config, arn string) (string, error) {
 	// sample: arn:aws:kinesis:region:account-id:stream/stream-name
 	matches := streamARNRE.FindStringSubmatch(arn)
 	if len(matches) != 4 {
 		return "", errInvalidStreamARN
 	}
-	cred, _ := conf.Credentials.Get()
-	if matches[1] != *conf.Region || matches[2] != cred.AccessKeyID {
-		return "", errInvalidSignature
-	}
+	// FIXME: AccountID check
+	// cred, _ := conf.Credentials.Retrieve(context.Background())
+	// if matches[1] != conf.Region || matches[2] != cred.AccountID {
+	// 	return "", errInvalidSignature
+	// }
 	return matches[3], nil
 }
 
-func newKinesisConsumer(ctx context.Context, conf *aws.Config, sourceConf *firehose.KinesisStreamSourceConfiguration, injectConf KinesisInjectedConf) (*kinesisConsumer, error) {
+func newKinesisConsumer(ctx context.Context, conf aws.Config, sourceConf *fhtypes.KinesisStreamSourceConfiguration, injectConf KinesisInjectedConf) (*kinesisConsumer, error) {
 	arn := sourceConf.KinesisStreamARN
 	if arn == nil {
-		return nil, awserr.New(firehose.ErrCodeInvalidArgumentException, "StreamARN not found", errInvalidStreamARN)
+		return nil, &fhtypes.InvalidArgumentException{Message: aws.String("StreamARN not found")}
 	}
 	streamName, err := getStreamNameFromARN(conf, *arn)
 	if err != nil {
-		return nil, awserr.New(firehose.ErrCodeInvalidArgumentException, "invalid StreamARN", err)
+		return nil, &fhtypes.InvalidArgumentException{Message: aws.String("invalid StreamARN")}
 	}
 	endpoint := injectConf.Endpoint
 	if endpoint == nil {
-		return nil, awserr.New(firehose.ErrCodeInvalidArgumentException, "KINESIS_STREAM_ENDPOINT_URL not found", errors.New("endpoint not found"))
+		return nil, &fhtypes.InvalidArgumentException{Message: aws.String("KINESIS_STREAM_ENDPOINT_URL not found")}
 	}
-	cli := kinesis.New(session.Must(session.NewSession(conf.Copy().WithEndpoint(*endpoint))))
-	out, err := cli.DescribeStreamWithContext(ctx, &kinesis.DescribeStreamInput{
+	cli := kinesis.NewFromConfig(conf, func(o *kinesis.Options) {
+		o.BaseEndpoint = endpoint
+	})
+	out, err := cli.DescribeStream(ctx, &kinesis.DescribeStreamInput{
 		StreamName: &streamName,
 	})
 	if err != nil {
-		return nil, awserr.New(firehose.ErrCodeServiceUnavailableException, "unable to connect Kinesis Streams", err)
+		return nil, &fhtypes.ServiceUnavailableException{Message: aws.String("unable to connect Kinesis Streams")}
 	}
 	desc := out.StreamDescription
-	if *desc.StreamStatus != "ACTIVE" {
-		return nil, awserr.New(firehose.ErrCodeServiceUnavailableException, "unable to connect Kinesis Streams", fmt.Errorf("stream status is %s", *desc.StreamStatus))
+	if desc.StreamStatus != types.StreamStatusActive {
+		return nil, &fhtypes.ServiceUnavailableException{Message: aws.String(fmt.Sprintf("stream status is %s", desc.StreamStatus))}
 	}
 	shardIter := map[string]string{}
 	for _, shard := range desc.Shards {
-		out, err := cli.GetShardIteratorWithContext(ctx, &kinesis.GetShardIteratorInput{
+		out, err := cli.GetShardIterator(ctx, &kinesis.GetShardIteratorInput{
 			StreamName:        &streamName,
 			ShardId:           shard.ShardId,
-			ShardIteratorType: aws.String("TRIM_HORIZON"),
+			ShardIteratorType: types.ShardIteratorTypeTrimHorizon,
 		})
 		if err != nil {
-			return nil, awserr.New(firehose.ErrCodeServiceUnavailableException, "failed to get shard iterator", err)
+			return nil, &fhtypes.ServiceUnavailableException{Message: aws.String("failed to get shard iterator")}
 		}
 		shardIter[*shard.ShardId] = *out.ShardIterator
 	}
@@ -78,7 +80,7 @@ func newKinesisConsumer(ctx context.Context, conf *aws.Config, sourceConf *fireh
 
 type kinesisConsumer struct {
 	streamName string
-	cli        *kinesis.Kinesis
+	cli        *kinesis.Client
 	shardIter  map[string]string
 }
 
@@ -89,20 +91,22 @@ func (c *kinesisConsumer) Run(ctx context.Context, recordCh chan *deliveryRecord
 	for shardID, iter := range c.shardIter {
 		go func(shardID, iter string) {
 			defer wg.Done()
+			currentIter := iter
 			for {
-				out, err := c.cli.GetRecordsWithContext(ctx, &kinesis.GetRecordsInput{
-					ShardIterator: &iter,
+				select {
+				case <-ctx.Done():
+					log.Debug().Str("shard_id", shardID).Msg("context canceled")
+					return
+				default:
+				}
+
+				out, err := c.cli.GetRecords(ctx, &kinesis.GetRecordsInput{
+					ShardIterator: &currentIter,
 				})
 				if err != nil {
-					switch ae, ok := err.(awserr.Error); {
-					case ok && ae.Code() == "RequestCanceled", err == context.Canceled:
-						log.Debug().Str("shard_id", shardID).Msg("context canceled")
-						return
-					default:
-						log.Debug().Str("shard_id", shardID).Err(err).Msg("GetRecord error")
-						time.Sleep(time.Second)
-						continue
-					}
+					log.Debug().Str("shard_id", shardID).Err(err).Msg("GetRecord error")
+					time.Sleep(time.Second)
+					continue
 				}
 				if l := len(out.Records); l > 0 {
 					log.Debug().Str("shard_id", shardID).Msgf("captured %d records", l)
@@ -110,7 +114,11 @@ func (c *kinesisConsumer) Run(ctx context.Context, recordCh chan *deliveryRecord
 				for _, record := range out.Records {
 					recordCh <- newDeliveryRecord(record.Data)
 				}
-				iter = *out.NextShardIterator
+				if out.NextShardIterator == nil {
+					log.Debug().Str("shard_id", shardID).Msg("NextShardIterator is nil, finishing subscription")
+					return
+				}
+				currentIter = *out.NextShardIterator
 				time.Sleep(time.Second)
 			}
 		}(shardID, iter)
